[RHS-145]: Fix user access to list courses endpoint
Details: Depending on whether a user is a User or an Admin, we fetch
courses from the DB in different ways. Either we fetch them using find
(Admins) or using aggregate (Users). However, results from an
aggregation are missing the .id field by default, which causes GraphQL
queries which expect that field to fail
(see [this issue](https://github.com/Automattic/mongoose/issues/8784)).
To fix this, we can pass the aggregation results through
MgCourse.hydrate, which will populate the expected fields.