Refactor ContentBlock typing
This is a large refactor with the primary goal of separating out the
state between different content types. This allows individual
renderers/editor modals to have strongly typed props, rather than having
to deal with a bunch of extraneous ones. It also allows for dynamically
selecting the component to render rather than having a giant switch
statement, which improves the legibility of the code.