Set up front-end structure to support more content types (#182)
* Add clientType property to ContentTypeEnum

* Refactor ContentBlock typing

This is a large refactor with the primary goal of separating out the
state between different content types. This allows individual renderers/editor modals to have strongly typed props, rather than having to deal with a bunch of extraneous ones. It also allows for dynamically selecting the component to render rather than having a giant switch statement, which improves the legibility of the code.

* Reorganize file structure to be more clear

Move files which are common to both the learner and admin side of things to the `common/` directory. Also move some ContentBlock types into their own file for organization.

* Course update mutation was broken on main

* PR feedback

1. Add comment explaining the purpose of the ContentBlockRenderer class.
2. Move all block information to a single table.
3. Add stronger type assertions to help catch certain classes of bugs.

* Mark all enum items as readonly

* Move string -> enum conversion logic to within the enum class